chore(sdk): move all data_types to same location
